+++ src/main/java/cart/application/repository/product/ProductRepository.java
@@ -1,6 +1,6 @@
-package cart.application.repository;
+package cart.application.repository.product;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 
 import java.util.List;
 import java.util.Optional;
