+++ src/main/java/cart/application/service/product/dto/ProductResultDto.java
@@ -1,6 +1,6 @@
 package cart.application.service.product.dto;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 
 public class ProductResultDto {
