--- src/main/java/cart/domain/Product.java
@@ -1,46 +1,0 @@
-package cart.domain;
-
-public class Product {
-    private Long id;
-    private String name;
-    private int price;
-    private String imageUrl;
-
-    public Product(String name, int price, String imageUrl) {
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
-
-    public Product(Long id, String name, int price, String imageUrl) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
-
-    public int calculatePriceBy(final int quantity) {
-        return quantity * price;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    public boolean isEqualId(final long productId) {
-        return id == productId;
-    }
-
-}
