+++ src/test/java/cart/application/service/ProductWriteServiceTest.java
@@ -1,6 +1,6 @@
 package cart.application.service;
 
-import cart.application.repository.ProductRepository;
+import cart.application.repository.product.ProductRepository;
 import cart.application.service.product.ProductWriteService;
 import cart.application.service.product.dto.ProductCreateDto;
 import org.junit.jupiter.api.DisplayName;
